requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
python-dateutil>=2.9.0.post0
